Move to babel support for ecmascript modules and flow.
